backend/requirements.txt
# FastAPI Backend Requirements - PostgreSQL Performance
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.6
httpx==0.26.0
beautifulsoup4==4.12.2
requests==2.31.0

# Templates e renderização
jinja2==3.1.3
aiofiles==23.2.1

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilitários
python-slugify==8.0.1
email-validator==2.1.0



